+++ src/components/MultiRangeSlider/index.ts
@@ -1,0 +1,4 @@
+import MultiRangeSlider from './MultiRangeSlider'
+
+export default MultiRangeSlider
+export * from './MultiRangeSlider'